Chest X-Ray Pneumonia Detection
In this project, images of chest X-rays is used to build and train a Deep Learning model using Convoluted Neural Networks (CNN) which can aid detect Pneumonia from chest X-rays. The model achieves a 95.6 % accuracy in pneumonia detection when tested on over 580 images, after running for 4 epochs. Visual plots have been presented to measure the change in the accuracy of the detections and change in loss over the 4 epochs.

 The dataset used is found on Kaggle at https://www.kaggle.com/datasets/paultimothymooney/chest-xray-pneumonia